refactor(internal/dbmodel/applicationoffer.go): removes data from the ApplicationOffer type that
Removes data from ApplicationOffer type that JIMM does not need to store. Instead the source of
truth for application offers should be Juju controllers.

It also allowed me to remove the update method for application offers and simplify the
AllWatcher so that we don't have to process deltas for offers.

A drive-by fix: we no longer require the user to have add-model permission to add a k8s cloud.