fix: flaky test in TestProxySocketsAdminFacade (#1474)
* fix: flaky test in TestProxySocketsAdminFacade

The test was incorrectly cancelling the context. When the proxySockets function encounters an error like a failure to connect to a Juju controller it returns with an error. Otherwise it returns a nil error when the context is cancelled. Here we were expecting the former for one of the tests cases but sometimes encountering the latter in a race condition.

* chore: remove expectProxyError bool from test